Updated SBE messages

+++ sbe/messages/src/messages/data_messages/start_data/sbe_decode.rs
@@ -17,7 +17,7 @@ pub fn decode_start_data_message(buffer: &[u8]) -> SbeResult<StartDataMessage> {
     assert_eq!(message_type, MessageType::StartData);
 
     let sbe_exchange_id = csg.exchange_id();
-    let exchange_id = ExchangeID::from(sbe_exchange_id as u8);
+    let exchange_id = ExchangeID::from(sbe_exchange_id as i32);
 
     let sbe_asset = csg.asset_id();
     let symbol = SymbolID::from(sbe_asset);
