+++ sbe/messages/src/messages/data_messages/stop_data/sbe_decode.rs
@@ -17,7 +17,7 @@ pub fn decode_stop_data_message(buffer: &[u8]) -> SbeResult<StopDataMessage> {
     assert_eq!(message_type, MessageType::StopData);
 
     let sbe_exchange_id = csg.exchange_id();
-    let exchange_id = ExchangeID::from(sbe_exchange_id as u8);
+    let exchange_id = ExchangeID::from(sbe_exchange_id as i32);
 
     let sbe_asset = csg.asset_id();
     let symbol = SymbolID::from(sbe_asset);
